chore

+++ src/ws.h
@@ -775,7 +775,7 @@ struct ws_server_params {
   const char *addr;
   uint16_t port;
   bool silent;              // logs server config to stdout if 0
-  size_t max_conns;        // Maximum connections the server is willing to accept.
+  unsigned max_conns;        // Maximum connections the server is willing to accept.
   size_t max_header_count; // Maximum number of headers to parse in the upgrade equest. (defaults to 32 and max is 512)
                              // Defaults to the system's limit for maximum open file descriptors.
   size_t max_buffered_bytes; // Maximum amount of websocket payload data to buffer before the connection
